initial wifi rework
